# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mawaeedak API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/upcoming:
    get:
      operationId: listUpcomingAppointments
      tags: [appointments]
      summary: List upcoming appointments
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Upcoming appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete appointment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Financial Events
  /financial-events:
    get:
      operationId: listFinancialEvents
      tags: [financial]
      summary: List financial events
      parameters:
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Financial events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialEvent"
    post:
      operationId: createFinancialEvent
      tags: [financial]
      summary: Create financial event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialEvent"
  /financial-events/{id}:
    patch:
      operationId: updateFinancialEvent
      tags: [financial]
      summary: Update financial event
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialEvent"
    delete:
      operationId: deleteFinancialEvent
      tags: [financial]
      summary: Delete financial event
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - name: city
          in: query
          schema: { type: string }
        - name: sector
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create job (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete job (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Prayer Times
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints (admin)
      responses:
        "200":
          description: Complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status / reply (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "404":
          description: Not found
    delete:
      operationId: deleteComplaint
      tags: [complaints]
      summary: Delete complaint (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Audit Logs
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, title, date, category, created_at]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        date: { type: string }
        time: { type: ["string", "null"] }
        category: { type: string }
        color: { type: ["string", "null"] }
        priority: { type: ["string", "null"] }
        reminder_enabled: { type: boolean }
        created_at: { type: string }
    AppointmentInput:
      type: object
      required: [title, date, category]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        date: { type: string }
        time: { type: string }
        category: { type: string }
        color: { type: string }
        priority: { type: string }
        reminder_enabled: { type: boolean }
    AppointmentUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        date: { type: string }
        time: { type: string }
        category: { type: string }
        color: { type: string }
        priority: { type: string }
        reminder_enabled: { type: boolean }
    FinancialEvent:
      type: object
      required: [id, name, type, next_date, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        next_date: { type: string }
        amount: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        confirmed: { type: boolean }
        source: { type: string }
        is_active: { type: boolean }
        reminder_days_before: { type: integer }
        created_at: { type: string }
    FinancialEventInput:
      type: object
      required: [name, type, next_date]
      properties:
        name: { type: string }
        type: { type: string }
        next_date: { type: string }
        amount: { type: number }
        confirmed: { type: boolean }
        source: { type: string }
        notes: { type: string }
        is_active: { type: boolean }
        reminder_days_before: { type: integer }
    FinancialEventUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        next_date: { type: string }
        amount: { type: number }
        confirmed: { type: boolean }
        source: { type: string }
        notes: { type: string }
        is_active: { type: boolean }
        reminder_days_before: { type: integer }
...
    Job:
      type: object
      required: [id, title, employer, sector, city, created_at]
      properties:
        id: { type: integer }
        title: { type: string }
        employer: { type: string }
        sector: { type: string }
        city: { type: string }
        description: { type: ["string", "null"] }
        apply_url: { type: ["string", "null"] }
        deadline: { type: ["string", "null"] }
        is_active: { type: boolean }
        created_at: { type: string }
    JobInput:
      type: object
      required: [title, employer, sector, city]
      properties:
        title: { type: string }
        employer: { type: string }
        sector: { type: string }
        city: { type: string }
        description: { type: string }
        apply_url: { type: string }
        deadline: { type: string }
        is_active: { type: boolean }
    JobUpdate:
      type: object
      properties:
        title: { type: string }
        employer: { type: string }
        sector: { type: string }
        city: { type: string }
        description: { type: string }
        apply_url: { type: string }
        deadline: { type: string }
        is_active: { type: boolean }
...
    Complaint:
      type: object
      required: [id, type, message, status, created_at]
      properties:
        id: { type: integer }
        type: { type: string }
        category: { type: ["string", "null"] }
        title: { type: ["string", "null"] }
        message: { type: string }
        contact: { type: ["string", "null"] }
        user_id: { type: ["string", "null"] }
        status: { type: string }
        admin_reply: { type: ["string", "null"] }
        created_at: { type: string }
        updated_at: { type: ["string", "null"] }
    ComplaintInput:
      type: object
      required: [type, message]
      properties:
        type: { type: string }
        category: { type: string }
        title: { type: string }
        message: { type: string }
        contact: { type: string }
        user_id: { type: string }
    ComplaintUpdate:
      type: object
      properties:
        status: { type: string }
        admin_reply: { type: string }
...